chore: introduce client-api-entity
