+++ libs/app-error/Cargo.toml
@@ -15,7 +15,6 @@ anyhow = "1.0.79"
 uuid = { version = "1.6.1", features = ["v4"] }
 sqlx = { version = "0.7", default-features = false, features = ["postgres", "json"], optional = true }
 validator = { version = "0.16", optional = true }
-rust-s3 = { version = "0.34.0-rc4", optional = true }
 url = { version = "2.5.0" }
 actix-web = { version = "4.4.1", optional = true }
 reqwest.workspace = true
@@ -28,7 +27,6 @@ appflowy-ai-client = { workspace = true, optional = true, default-features = fal
 default = []
 sqlx_error = ["sqlx"]
 validation_error = ["validator"]
-s3_error = ["rust-s3"]
 actix_web_error = ["actix-web"]
 tokio_error = ["tokio"]
 gotrue_error = []
