+++ libs/client-api-entity/src/lib.rs
@@ -1,0 +1,9 @@
+pub use collab_entity::*;
+pub use collab_rt_entity::user::*;
+pub use database_entity::dto::*;
+pub use database_entity::file_dto::*;
+pub use gotrue_entity::dto::*;
+pub use shared_entity::dto::*;
+
+#[cfg(feature = "file_util")]
+pub use infra::file_util;
