+++ libs/client-api/src/http_chat.rs
@@ -1,7 +1,7 @@
 use crate::http::log_request_id;
 use crate::Client;
 use bytes::Bytes;
-use database_entity::dto::{
+use client_api_entity::{
   ChatMessage, CreateAnswerMessageParams, CreateChatMessageParams, CreateChatParams, MessageCursor,
   RepeatedChatMessage, UpdateChatMessageContentParams,
 };
