+++ libs/client-api/src/lib.rs
@@ -36,11 +36,7 @@ pub mod error {
 
 // Export all dto entities that will be used in the frontend application
 pub mod entity {
-  pub use collab_rt_entity::user::*;
-  pub use database_entity::dto::*;
-  pub use database_entity::file_dto::*;
-  pub use gotrue_entity::dto::*;
-  pub use shared_entity::dto::*;
+  pub use client_api_entity::*;
 }
 
 #[cfg(feature = "template")]
