+++ libs/database/Cargo.toml
@@ -10,7 +10,7 @@ collab = { workspace = true }
 collab-entity = { workspace = true }
 validator = { version = "0.16", features = ["validator_derive", "derive"] }
 database-entity.workspace = true
-app-error = { workspace = true, features = ["sqlx_error", "validation_error", "s3_error"] }
+app-error = { workspace = true, features = ["sqlx_error", "validation_error"] }
 
 tokio = { workspace = true, features = ["sync"] }
 async-trait = "0.1.77"
