+++ libs/infra/Cargo.toml
@@ -6,8 +6,14 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-reqwest = { workspace = true }
+reqwest = { workspace = true, optional = true }
 anyhow = "1.0.79"
 serde.workspace = true
 serde_json.workspace = true
 tracing.workspace = true
+bytes = { workspace = true, optional = true }
+tokio = { workspace = true, optional = true }
+
+[features]
+file_util = ["bytes", "tokio/fs"]
+request_util = ["reqwest"]
