+++ libs/infra/src/lib.rs
@@ -1,2 +1,6 @@
 pub mod env_util;
+
+#[cfg(feature = "file_util")]
+pub mod file_util;
+#[cfg(feature = "request_util")]
 pub mod reqwest;
